+++ docs/release_notes.rst
@@ -7,6 +7,7 @@ Version 0.3.4 (under development)
 
 * Use multi certificates with PaypalEncryptedPaymentsForm
 * Fixed issue #166 - regression from 0.2.7 when using ``USE_TZ=False``
+* Django 1.11 compatibility.
 
 Version 0.3.3
 -------------